Tests changed
Added some tests
